Split scheduler tests into scheduler/tests/{support,unit,execution}.rs
